Revert "chore: Update Spotify integration in navbar"
