+++ src/components/Navbar.astro
@@ -1,6 +1,6 @@
 ---
 import { navItems } from "../utils/navigation";
-import Badge from "@components/Spotify/Badge.astro";
+import { SpotifyNowPlaying } from "@components/Spotify/Badge";
 
 type Props = {
   display?: boolean;
@@ -30,6 +30,6 @@ const { display } = Astro.props;
         </li>
       ))
     }
-    {display ?? <Badge />}
+    {display ?? <SpotifyNowPlaying client:only="react" />}
   </ul>
 </nav>
