--- src/components/Spotify/Badge.astro
@@ -1,20 +1,0 @@
----
-import { getCurrentTrack } from "@utils/spotify";
-import { NowPlaying } from "./NowPlaying";
-import { NotPlaying } from "./NotPlaying";
-
-export const prerender = false;
-
-const response = await getCurrentTrack();
-const currentTrack = await response.json();
----
-
-<li class="hidden md:block">
-  {
-    currentTrack.isPlaying ? (
-      <NowPlaying props={currentTrack} client:only="react" />
-    ) : (
-      <NotPlaying />
-    )
-  }
-</li>
