+++ src/components/Spotify/NowPlaying.tsx
@@ -30,8 +30,8 @@ export const NowPlaying = ({ props }: { props: NowPlayingTrackResponse }) => {
         width="24"
         height="24"
         xmlns="http://www.w3.org/2000/svg"
-        fillRule="evenodd"
-        clipRule="evenodd"
+        fill-rule="evenodd"
+        clip-rule="evenodd"
         className="heartbeat"
       >
         <path
